[#493] node/config: Export useful functions into a separate test package
