+++ cmd/neofs-node/config/calls_test.go
@@ -4,11 +4,12 @@ import (
 	"testing"
 
 	"github.com/nspcc-dev/neofs-node/cmd/neofs-node/config"
+	configtest "github.com/nspcc-dev/neofs-node/cmd/neofs-node/config/test"
 	"github.com/stretchr/testify/require"
 )
 
 func TestConfigCommon(t *testing.T) {
-	forEachFileType("test/config", func(c *config.Config) {
+	configtest.ForEachFileType("test/config", func(c *config.Config) {
 		val := c.Value("value")
 		require.NotNil(t, val)
