+++ cmd/neofs-node/config/cast_test.go
@@ -4,11 +4,12 @@ import (
 	"testing"
 
 	"github.com/nspcc-dev/neofs-node/cmd/neofs-node/config"
+	configtest "github.com/nspcc-dev/neofs-node/cmd/neofs-node/config/test"
 	"github.com/stretchr/testify/require"
 )
 
 func TestStringSlice(t *testing.T) {
-	forEachFileType("test/config", func(c *config.Config) {
+	configtest.ForEachFileType("test/config", func(c *config.Config) {
 		cStringSlice := c.Sub("string_slice")
 
 		val := config.StringSlice(cStringSlice, "empty")
@@ -30,7 +31,7 @@ func TestStringSlice(t *testing.T) {
 }
 
 func TestString(t *testing.T) {
-	forEachFileType("test/config", func(c *config.Config) {
+	configtest.ForEachFileType("test/config", func(c *config.Config) {
 		c = c.Sub("string")
 
 		val := config.String(c, "correct")
